updated for deletion

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../models/Section");
 const Course = require("../models/Course");
+const SubSection = require("../models/SubSection") || mongoose.model('SubSection');
 // CREATE a new section
 exports.createSection = async (req, res) => {
 	try {
@@ -74,16 +75,43 @@ exports.updateSection = async (req, res) => {
 };
 
 // DELETE a section
+// Delete a section and update the associated course
 exports.deleteSection = async (req, res) => {
 	try {
-		// HW : fetch with req.params
+
 		const { sectionId, courseId } = req.body;
+		await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+		const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if (!section) {
+			return res.status(404).json({
+				success: false,
+				message: "Section not Found",
+			})
+		}
+
+		//delete sub section
+		await SubSection.deleteMany({ _id: { $in: section.subSection } });
+
 		await Section.findByIdAndDelete(sectionId);
-		// HW >>
-		// TODO : Course ko bhi update karo
+
+		//find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path: "courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+			.exec();
+
 		res.status(200).json({
 			success: true,
 			message: "Section deleted",
+			data: course
 		});
 	} catch (error) {
 		console.error("Error deleting section:", error);
@@ -92,4 +120,4 @@ exports.deleteSection = async (req, res) => {
 			message: "Internal server error",
 		});
 	}
-};
+};
